requirements.txt
streamlit
pandas
numpy
plotly
akshare
pdfplumber
openai
python-dotenv
langchain
langchain-openai
jieba
faiss-cpu
rank-bm25
fastapi
uvicorn[standard]
